refactor: Extract PagesLayoutContent and simplify PagesLayout
- Moved sidebar and layout logic from PagesLayout to new PagesLayoutContent
- Wrapped PagesLayoutContent with Suspense and loading fallback in PagesLayout
- Simplified Sidebar toggle handlers by removing mutual exclusivity logic
- Removed redundant localStorage logic from Sidebar item click handler
- Improved code separation and readability by splitting concerns into components
